+++ vice_driver/__init__.py
@@ -61,4 +61,4 @@ __all__ = [
     "write_png",
 ]
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"